Change some nits
Made sure some functions are documented while other private functions
are excluded (the ones that shouldn't be called).

While writing more docs for the `db` module, I noticed that there is a
big need for an `update` function that updates a single (or multiple)
values in the database. It might be a bit difficult, but it should be
done to avoid the entire charade of pulling and pushing and overwriting.