Remove factory from testing
Beforehand, factory tests only tested if the function returned an
object of the same type. Now that we have mypy, this is obsolete. Apart
from this, we have found it increasingly difficult to cater to both the
needs of the application and the needs of the tests (mocks vs real
things). Thus we have decided to remove factory testing.

We update `.coverageac` accordingly.