cleanup: rm docstrings req and enforce PEP 420
We cannot completely remove the `__init__.py` files from the `tests/`
directory because in [the section below this one][pytest-doc], it says
that even if our code does namespace packages correctly (PEP 420), we
still need to add `__init__.py` files to the tests for them to be
discovered properly. What a bummer.

[pytest-doc]: https://docs.pytest.org/en/latest/goodpractices.html#tests-as-part-of-application-code